+++ packages/code-du-travail-frontend/src/outils/DureePreavisDemission/stepReducer.ts
@@ -1,11 +1,11 @@
-import data from "@cdt/data...simulateurs/preavis-licenciement.data.json";
+import data from "@cdt/data...simulateurs/preavis-demission.data.json";
 
+import { MatomoActionEvent } from "../../lib";
 import { isNotYetProcessed } from "../common/situations.utils";
-import { StepInformations } from "./steps/Informations";
+import { StepInformations } from "../common/StepInformations";
 import { StepInfoCCn } from "./steps/InfosCCn";
 import { StepIntro } from "./steps/Introduction";
 import { StepResult } from "./steps/Result";
-import { StepStatus } from "./steps/Status";
 
 export const initialState = {
   stepIndex: 0,
@@ -16,18 +16,18 @@ export const initialState = {
       name: "intro",
     },
     {
-      component: StepStatus,
-      label: "Situation du salarié",
-      name: "situation",
-    },
-    {
       component: StepInfoCCn,
       label: "Convention collective",
       name: "info_cc",
     },
     {
       component: StepInformations,
-      label: "Informations complémentaires",
+      componentProps: {
+        actionEvent: MatomoActionEvent.PREAVIS_DEMISSION,
+      },
+      hasNoMarginBottom: true,
+      isForm: true,
+      label: "Informations",
       name: "infos",
       skip: (values) =>
         !values.ccn ||
