+++ packages/code-du-travail-frontend/src/outils/DureePreavisLicenciement/stepReducer.ts
@@ -1,10 +1,12 @@
-import data from "@cdt/data...simulateurs/preavis-demission.data.json";
+import data from "@cdt/data...simulateurs/preavis-licenciement.data.json";
 
+import { MatomoActionEvent } from "../../lib";
 import { isNotYetProcessed } from "../common/situations.utils";
-import { StepInformations } from "./steps/Informations";
+import { StepInformations } from "../common/StepInformations";
 import { StepInfoCCn } from "./steps/InfosCCn";
 import { StepIntro } from "./steps/Introduction";
 import { StepResult } from "./steps/Result";
+import { StepStatus } from "./steps/Status";
 
 export const initialState = {
   stepIndex: 0,
@@ -15,15 +17,21 @@ export const initialState = {
       name: "intro",
     },
     {
+      component: StepStatus,
+      label: "Situation du salarié",
+      name: "situation",
+    },
+    {
       component: StepInfoCCn,
       label: "Convention collective",
       name: "info_cc",
     },
     {
       component: StepInformations,
-      hasNoMarginBottom: true,
-      isForm: true,
-      label: "Informations",
+      componentProps: {
+        actionEvent: MatomoActionEvent.PREAVIS_LICENCIEMENT,
+      },
+      label: "Informations complémentaires",
       name: "infos",
       skip: (values) =>
         !values.ccn ||
