--- packages/code-du-travail-frontend/src/outils/DureePreavisLicenciement/steps/__tests__/Informations.test.js
@@ -1,18 +1,0 @@
-import { renderForm } from "../../../../../test/renderForm";
-import { StepInformations } from "../Informations";
-
-describe("<StepInformations />", () => {
-  it("should render catégorie", () => {
-    const { container } = renderForm(StepInformations, {
-      ccn: { num: 44 },
-    });
-    expect(container).toMatchSnapshot();
-  });
-  it("should render groupe", () => {
-    const { container } = renderForm(StepInformations, {
-      ccn: { num: 44 },
-      criteria: { "catégorie professionnelle": "23| Agents de maîtrise" },
-    });
-    expect(container).toMatchSnapshot();
-  });
-});
